requirements.txt
gradio
websocket-client 
requests-toolbelt
